begin working on a refactored API
