+++ api/pyproject.toml
@@ -1,0 +1,63 @@
+[tool.poetry]
+name = "serge"
+description = "Serge API package"
+version = "0.1.0"
+license = "MIT"
+authors = [ 
+    "Nathan Sarrazin <contact@nsarrazin.com>"
+]
+packages = [
+    { include = "*", from = "src" }
+]
+
+homepage = "https://serge.chat/"
+repository = "https://github.com/nsarrazin/serge"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.dependencies]
+python=">=3.10,<4.0"
+asyncio = "^3.4.3"
+packaging = "^23.0"
+pydantic = "^1.10.7"
+pymongo = "^4.3.3"
+python-dotenv = "^1.0.0"
+python-multipart = "^0.0.6"
+pyyaml = "^6.0"
+rfc3986 = "^2.0.0"
+sentencepiece = "^0.1.97"
+sniffio = "^1.3.0"
+sse-starlette = "^1.3.3"
+starlette = "^0.26.1"
+toml = "^0.10.2"
+tqdm = "^4.65.0"
+typing-extensions = "^4.5.0"
+ujson = "^5.7.0"
+urllib3 = "^1.26.15"
+uvicorn = "^0.21.1"
+uvloop = "^0.17.0"
+watchfiles = "^0.19.0"
+websockets = "^10.4"
+anyio = "^3.6.2"
+certifi = "^2022.12.7"
+charset-normalizer = "^3.1.0"
+click = "^8.1.3"
+email-validator = "^1.3.1"
+fastapi = "^0.95.0"
+filelock = "^3.10.7"
+h11 = "^0.14.0"
+httpcore = "^0.17.0"
+httptools = "^0.5.0"
+huggingface-hub = "^0.13.3"
+idna = "^3.4"
+itsdangerous = "^2.1.2"
+jinja2 = "^3.1.2"
+markupsafe = "^2.1.2"
+motor = "^3.1.1"
+orjson = "^3.8.8"
+beanie = "^1.17.0"
+dnspython = "^2.3.0"
+lazy-model = "^0.0.5"
+requests = "^2.28.2"
