+++ api/src/routers/__init__.py
@@ -1,0 +1,2 @@
+from .chat import chat_router
+from .model import model_router
