+++ api/src/routers/model.py
@@ -1,0 +1,6 @@
+from fastapi import APIRouter
+
+chat_router = APIRouter(
+    prefix="/model",
+    tags=["model"],
+)
