+++ api/src/utils/generate.py
